Various review related changes - Changed EnvironmentControl.Exit to accept ErrorCode Enum instead of int Related to Issue [WTI-1060]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@8047 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: 0419d307421da40b733e5ab4aefef17b648baad5
Former-commit-id: 1d72404e8e30e3c42e2e44ab1f340a32011cab97

+++ Application/Ringtoets/src/Application.Ringtoets.MigrationConsole/RingtoetsMigrationTool.cs
@@ -74,7 +74,7 @@ namespace Application.Ringtoets.MigrationConsole
 
         private static void Exit(ErrorCode errorCode)
         {
-            EnvironmentControl.Instance.Exit((int) errorCode);
+            EnvironmentControl.Instance.Exit(errorCode);
         }
 
         private static void DisplayException(Exception exception)
